Atualizando

+++ app/src/views/Admin/Users/index.tsx
@@ -2,6 +2,9 @@ import { useCallback, useEffect, useState } from 'react'
 import { AdminProps } from '..'
 import { CardSettings } from '../../../components/Cards/CardSettings'
 import { Company } from '../Companys'
+import { UserCreate, UserData } from '../../../components/Modal/UserCreate'
+import { User } from '../../../context/AuthContext'
+import { UserEditModal } from '../../../components/Modal/UserEditModal'
 
 type lUser = {
   relation: null | Company
@@ -13,7 +16,9 @@ type lUser = {
 export const Users = ({ permissions, apiclient }: AdminProps) => {
   const [users, setUsers] = useState<User[]>([])
   const [companys, setCompanys] = useState<Company[]>([])
-  const [isModalOpen, setIsModalOpen] = useState(false)
+  const [isModalCreateUserOpen, setIsModalCreateUserOpen] = useState(false)
+  const [isModalEditUserOpen, setIsModalEditUserOpen] = useState(false)
+  const [userData, setUserData] = useState<User | null>(null)
   const loadCompanys = useCallback(async () => {
     const companys: Company[] = await apiclient.get('/company')
     setCompanys(companys)
@@ -22,14 +27,41 @@ export const Users = ({ permissions, apiclient }: AdminProps) => {
     const users: lUser[] = await apiclient.get('/user')
     setUsers(users)
   }, [apiclient])
-  const handleConfirm = async (userData: UserData) => {
-    await apiclient.post('/user', userData)
-    loadUsers()
-    setIsModalOpen(false)
+  const handleCreateUserConfirm = async (userData: UserData) => {
+    if (!userData.company) alert('Choose a company')
+    if (!userData.name) alert('Name is required')
+    if (!userData.username) alert('Username is required')
+    if (!userData.password) alert('Password is required')
+    try {
+      await apiclient.post('/user', userData)
+      loadUsers()
+      setIsModalCreateUserOpen(false)
+    } catch (error) {
+      console.log(error)
+    }
   }
-  const handleDelete = async (id: string) => {
+  const handleDelete = async (id: string | undefined) => {
+    if (!id) return
+    if (!window.confirm('Tem certeza que deseja excluir este usuário?')) return
+
     await apiclient.delete(`/user/${id}`)
     loadUsers()
+  }
+  const handleEdit = async (user: User) => {
+    setUserData(user)
+    setIsModalEditUserOpen(true)
+  }
+  const handleUpdate = async (userData: User) => {
+    try {
+      delete userData.relation
+      delete userData.created_at
+
+      await apiclient.put(`/user/${userData._id}`, userData)
+      loadUsers()
+      setIsModalEditUserOpen(false)
+    } catch (error) {
+      console.log(error)
+    }
   }
   useEffect(() => {
     loadCompanys()
@@ -37,7 +69,28 @@ export const Users = ({ permissions, apiclient }: AdminProps) => {
   }, [loadCompanys, loadUsers])
   return (
     <div className='relative pb-32 pt-12'>
-      <CardSettings users={users} />
+      <CardSettings
+        permissions={permissions}
+        users={users}
+        createUser={() => setIsModalCreateUserOpen(true)}
+        onDelete={handleDelete}
+        onEdit={handleEdit}
+      />
+      <UserCreate
+        isModalOpen={isModalCreateUserOpen}
+        handleClose={() => setIsModalCreateUserOpen(false)}
+        handleConfirm={handleCreateUserConfirm}
+        companys={companys}
+      />
+      <UserEditModal
+        isModalOpen={isModalEditUserOpen}
+        handleSave={handleUpdate}
+        handleClose={() => {
+          setIsModalEditUserOpen(false)
+          setUserData(null)
+        }}
+        userData={userData}
+      />
     </div>
   )
 }
